+++ src/index.ts
@@ -3,3 +3,4 @@ export { default } from "./lib/remarkWikiLink";
 export { syntax } from "./lib/syntax";
 export { fromMarkdown } from "./lib/fromMarkdown";
 export { toMarkdown } from "./lib/toMarkdown";
+export { defaultUrlResolver } from "./utils/index";