+++ src/utils/const.ts
@@ -1,0 +1,2 @@
+// export const WIKI_LINK_TARGET_PATTERN = /^(.*?)(?:#(.*))?$/u;
+export const WIKI_LINK_TARGET_PATTERN = /^(.*?)(#.*)?$/u;